+++ next/src/App.tsx
@@ -6,6 +6,7 @@ import i18next from '@/lib/i18n'
 
 import ErrorComponent from './components/layout/ErrorComponent'
 import Layout from './components/layout/Layout'
+import WaitingComponent from './components/layout/WaitingComponent'
 import { routeTree } from './routeTree.gen'
 
 const queryClient = new QueryClient()
@@ -26,6 +27,7 @@ const router = createRouter({
       />
     </Layout>
   ),
+  defaultPendingComponent: WaitingComponent,
 })
 
 // Register the router instance for type safety