+++ next/src/api/personalization.ts
@@ -8,8 +8,9 @@ import {
   UploadMessage,
 } from '@/models/personalizationQuestionnaire'
 
-import { instancePersonalization } from './instancePersonalization'
-import { getFileName, openDocument } from './utils/personalization'
+import { instancePersonalization } from './instancePersonalization';
+import { getFileName } from './utils/personalization';
+import { openDocument } from '@/utils/files';
 
 export const personalizationKeys = {
   base: (poguesId: string) => ['personalization', 'base', poguesId] as const,