+++ next/src/components/personalization/overview/PersonalizationOverview.tsx
@@ -8,17 +8,17 @@ import {
   deleteQuestionnaireData,
   personalizationKeys,
 } from '@/api/personalization'
-import { openParsedCsv, openParsedJson } from '@/api/utils/personalization'
 import PersonalizationContentTile from '@/components/personalization/overview/PersonalisationContentTile'
 import Button, { ButtonStyle } from '@/components/ui/Button'
 import ButtonLink from '@/components/ui/ButtonLink'
+import CsvViewerTable from '@/components/ui/CsvViewerTable'
 import DialogButton from '@/components/ui/DialogButton'
 import {
   InterrogationModeDataResponse,
   PersonalizationQuestionnaire,
 } from '@/models/personalizationQuestionnaire'
+import { openParsedCsv, openParsedJson } from '@/utils/files'
 
-import CsvViewerTable from '../form/CsvViewerTable'
 import JsonViewer from '../form/JsonViewer'
 import PersonalizationCheckPanel from './PersonalizationCheckPanel'
 import PersonalisationTile from './PersonalizationTile'